- Philip Wyett eliminated signed/unsigned mismatch warning in   function csString::SubString ().
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@14296 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csutil/csstring.cpp
@@ -183,7 +183,7 @@ csString &csString::Append (const char *iStr, size_t iCount)
 
 void csString::SubString (csString& sub, size_t x, size_t len)
 {
-  int y = x + len;
+  size_t y = x + len;
 
   if (x < 0 || x >= Size || y < 0 || y > Size || y <= x) 
   {
